requirements.txt
streamlit
openai
python-dotenv
azure-ai-formrecognizer
transformers
azure-search-documents